[ADD] .env variables

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { TaskRepository } from './tasks.repository';
 import { TaskStatus } from './task-status.emun';
 import { Task } from './tasks.entity';
 import { User } from '../auth/user.entity';
+import { use } from 'passport';
 
 @Injectable()
 export class TasksService {
@@ -18,30 +19,30 @@ export class TasksService {
     return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number) {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const find = await this.taskRepository.findOne({ where: { id, userId: user.id }});
 
-    if (!found) {
+    if (!find) {
       throw new NotFoundException(`Task with ID '${id}' not found!`);
     }
 
-    return found;
+    return find;
   }
 
   async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
     return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID '${id}' not found!`);
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
